+++ Packages/com.unity.render-pipelines.universal@14.0.8/Runtime/Passes/PostProcessPass.cs
@@ -489,6 +489,17 @@ namespace UnityEngine.Rendering.Universal
                 }
             }
 
+            //************** CUSTOM ADD START ***************//
+            if (activeVolumeRenderers != null)
+            {
+                for (int i = 0; i < activeVolumeRenderers.Count; i++)
+                {
+                    activeVolumeRenderers[i].Render(cmd, GetSource(), GetDestination());
+                    Swap(ref renderer);
+                }
+            }
+            //*************** CUSTOM ADD END ****************//
+            
             // Anti-aliasing
             if (useSubPixeMorpAA)
             {